refactor: (#164) 활동 도메인 응용서비스를 도메인 서비스로 이관한다
