+++ src/main/java/spring/backend/activity/application/UserActivitySelectService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import spring.backend.activity.domain.entity.Activity;
 import spring.backend.activity.domain.repository.ActivityRepository;
+import spring.backend.activity.domain.service.FinishActivityAutoService;
 import spring.backend.activity.dto.request.UserActivitySelectRequest;
 import spring.backend.activity.dto.response.UserActivitySelectResponse;
 import spring.backend.activity.exception.ActivityErrorCode;
